loads and loads of modifications concerning mainly the organization of the code and other things you might check

+++ Project_Code_S2019/Restaurant/Events/ArrivalEvent.cpp
@@ -8,7 +8,7 @@ ArrivalEvent::ArrivalEvent(int eTime, int oID, ORD_TYPE oType, REGION reg):Event
 	OrdRegion = reg;
 }
 
-ArrivalEvent::ArrivalEvent(int eTime, int oID, ORD_TYPE oType, REGION reg, int dist, double mon):Event(eTime, oID)
+ArrivalEvent::ArrivalEvent(int eTime, int oID, ORD_TYPE oType, REGION reg, double dist, double mon):Event(eTime, oID)
 {
 	OrdType = oType;
 	OrdRegion = reg;
@@ -23,7 +23,8 @@ void ArrivalEvent::Execute(Restaurant* pRest)
 {
 	//This function should create an order and fills its data 
 	// Then adds it to normal, frozen, or VIP order lists that you will create in phase1
-	Order* pOrd = new Order(OrderID,OrdType,OrdRegion, OrdDistance, OrdMoney);
+	
+	Order* pOrd = new Order(OrderID, OrdType, OrdRegion, OrdDistance, OrdMoney);
 	//filling the order to the lists in its region
 	if(OrdType == TYPE_NRM)
 		pRest->GetRegion(OrdRegion)->InsertNOrder(pOrd);
