+++ Project_Code_S2019/Restaurant/Events/ArrivalEvent.h
@@ -7,16 +7,13 @@
 //class for the arrival event
 class ArrivalEvent: public Event
 {
-	//info about the order ralted to arrival event
-	int OrdDistance;	//order distance
+	double OrdDistance;  	//order distance
 	ORD_TYPE OrdType;		//order type: Normal, Frozen, VIP
-	REGION OrdRegion;  //Region of this order	                
-	double OrdMoney;	//Total order money
+	REGION OrdRegion;       //Region of this order	                
+	double OrdMoney;	    //Total order money
 public:
 	ArrivalEvent(int eTime, int oID, ORD_TYPE oType, REGION reg);
-	//Add more constructors if needed
-	ArrivalEvent(int eTime, int oID, ORD_TYPE oType, REGION reg, int dist, double mon);
-	
+	ArrivalEvent(int eTime, int oID, ORD_TYPE oType, REGION reg, double dist, double mon);
 	virtual void Execute(Restaurant *pRest);	//override execute function
 
 };
